Enum cursive::align::VAlign

pub enum VAlign {
    Top,
    Center,
    Bottom,
}

Vertical alignment

Variants
Top

Place the element at the top of available space

Center

Place the element vertically in the center of available space

Bottom

Place the element at the bottom of available space

Methods

impl VAlign

pub fn get_offset(&self, content: usize, container: usize) -> usize

Returns the offset required to position a view.

When drawing a view with size content when the available size is container, printing at the resulting offset will align the view as desired.