Module cursive::direction

Direction-related structures.

This module defines two main concepts: [Orientation] and [Direction].

Orientation

Orientation is a simple enum that can take two values: Horizontal or Vertical.

Direction

Direction is a bit more complex, and can be of two kinds:

Absolute direction: left, up, right, or down

Relative direction: front or back. Its actual direction depends on the orientation.

Enums

Absolute
Absolute direction (up, down, left, right).

Direction
Represents a direction, either absolute or orientation-dependent.

Orientation
Describes a vertical or horizontal orientation for a view.

Relative
Direction relative to an orientation.