Struct cursive::event::Callback

pub struct Callback(_);

Callback is a function that can be triggered by an event. It has a mutable access to the cursive root.

Methods

impl Callback

pub fn from_fn<F: Fn(&mut Cursive) + 'static>(f: F) -> Self

Wraps the given function into a Callback object.