Module cursive::prelude

Commonly used imports, conveniently grouped.

To easily import a bunch of commonly-used definitions, import this bundle:

use cursive::prelude::*;

Reexports

pub use ::{Cursive};

pub use event::{Event, Key};

pub use view::{Selector, View};

pub use vec::Vec2;

pub use menu::MenuTree;

Structs

BoxView
Wrapper around another view, with a controlled size.

Button
Simple text label with a callback when is pressed.

Checkbox
Checkable box.

Dialog
Popup-like view with a main content, and optional buttons under it.

EditView
Input box where the user can enter and edit text.

FullView
Simple wrapper view that asks for all the space it can get.

IdView
Wrapper view that allows to select its content with a fixed string id.

KeyEventView
A simple wrapper view that catches some ignored event from its child.

LinearLayout
Arranges its children linearly according to its orientation.

ListView
Displays a scrollable list of elements.

Panel
Draws a border around a wrapped view.

Printer
Convenient interface to draw on a subset of the screen.

ProgressBar
Animated bar showing a progress value.

SelectView
View to select an item among a list.

TextArea
Multi-lines text editor.

TextView
A simple view showing a fixed text

Traits

Boxable
Makes a view wrappable in a BoxView.

Identifiable
Makes a view wrappable in an IdView.

With
Generic trait to enable chainable API