Struct cursive::printer::Printer

pub struct Printer {
    pub offset: Vec2,
    pub size: Vec2,
    pub focused: bool,
    pub theme: Theme,
}

Convenient interface to draw on a subset of the screen.

Fields
offset: Vec2

Offset into the window this printer should start drawing at.

size: Vec2

Size of the area we are allowed to draw on.

focused: bool

Whether the view to draw is currently focused or not.

theme: Theme

Currently used theme

Methods

impl Printer

fn new<T: ToVec2>(size: T, theme: Theme) -> Self

Creates a new printer on the given window.

fn print<S: ToVec2>(&self, pos: S, text: &str)

Prints some text at the given position relative to the window.

fn print_vline<T: ToVec2>(&self, start: T, len: usize, c: chtype)

Prints a vertical line using the given character.

fn print_hline<T: ToVec2>(&self, start: T, len: usize, c: chtype)

Prints a horizontal line using the given character.

fn with_color<'a, F>(&'a self, c: ColorPair, f: F) where F: Fn(&Printer)

Call the given closure with a colored printer, that will apply the given color on prints.

Examples

printer.with_color(theme::ColorPair::Highlight, |printer| {
    printer.print((0,0), "This text is highlighted!");
});

fn with_style<'a, F>(&'a self, style: attr_t, f: F) where F: Fn(&Printer)

Same as with_color, but apply a ncurses style instead, like ncurses::A_BOLD() or ncurses::A_REVERSE().

Will probably use a cursive enum some day.

fn print_box<T: ToVec2>(&self, start: T, size: T)

Prints a rectangular box.

Examples

printer.print_box((0,0), (6,4));

fn sub_printer<S: ToVec2>(&self, offset: S, size: S, focused: bool) -> Printer

Returns a printer on a subset of this one's area.