Struct cursive::Menubar

pub struct Menubar {
    pub menus: Vec<(String, Rc<MenuTree>)>,
    pub autohide: bool,
    pub focus: usize,
    // some fields omitted
}

Shows a single-line list of items, with pop-up menus when one is selected.

The Cursive root already includes a menubar that you just needs to configure.

Fields
menus: Vec<(String, Rc<MenuTree>)> autohide: bool focus: usize
Methods

impl Menubar

fn new() -> Self

Creates a new, empty menubar.

fn take_focus(&mut self)

Takes the focus.

TODO: impl View

fn receive_events(&self) -> bool

True if we should be receiving events.

fn visible(&self) -> bool

Returns true if we should be drawn.

fn add(&mut self, title: &str, menu: MenuTree) -> &mut Self

Adds a new item to the menubar.

The item will use the given title, and on selection, will open a popup-menu with the given menu tree.

fn draw(&mut self, printer: &Printer)

Draws the view.

TODO: impl View

fn on_event(&mut self, event: Event) -> Option<Callback>

Reacts to event.

TODO: impl View