Struct cursive::XY

pub struct XY<T> {
    pub x: T,
    pub y: T,
}

A generic structure with a value for each axis.

Fields
x: T

X-axis value

y: T

Y-axis value

Methods

impl XY<usize>

fn max<A: Into<Vec2>, B: Into<Vec2>>(a: A, b: B) -> Self

Returns a new Vec2 that is a maximum per coordinate.

fn min<A: Into<Vec2>, B: Into<Vec2>>(a: A, b: B) -> Self

Returns a new Vec2 that is no larger than any input in both dimensions.

fn or_min<T: Into<Vec2>>(self, other: T) -> Self

Returns the minimum of self and other.

fn or_max<T: Into<Vec2>>(self, other: T) -> Self

Returns the maximum of self and other.

fn keep_x(&self) -> Self

Returns a vector with the X component of self, and y=0.

fn keep_y(&self) -> Self

Returns a vector with the Y component of self, and x=0.

fn zero() -> Self

Alias for Vec2::new(0,0).

fn stack_vertical(&self, other: &Vec2) -> Vec2

Returns (max(self.x,other.x), self.y+other.y)

fn stack_horizontal(&self, other: &Vec2) -> Vec2

Returns (self.x+other.x, max(self.y,other.y))

fn fits_in<T: Into<Vec2>>(&self, other: T) -> bool

Returns true if self could fit inside other.

Shortcut for self.x <= other.x && self.y <= other.y.

fn with_axis(&self, o: Orientation, value: usize) -> Self

Returns a new Vec2 with the axis o set to value.

fn with_axis_from(&self, o: Orientation, other: &Vec2) -> Self

Returns a new Vec2 with the axis o set to the value from other.

fn set_axis_from(&mut self, o: Orientation, other: &Vec2)

Sets the axis o on self to the value from other.

impl<T> XY<T>

fn new(x: T, y: T) -> Self

Creates a new XY from the given values.

fn map<U, F: Fn(T) -> U>(self, f: F) -> XY<U>

Creates a new XY by applying f to x and y.

fn pair(self) -> (T, T)

Destructure self into a pair.

fn as_ref(&self) -> XY<&T>

Return a XY with references to this one's values.

fn iter(&self) -> Chain<Once<&T>, Once<&T>>

Creates an iterator that returns references to x, then y.

fn get(&self, o: Orientation) -> &T

Returns a reference to the value on the given axis.

fn zip<U>(self, other: XY<U>) -> XY<(T, U)>

Returns a new XY of tuples made by zipping self and other.

fn zip_map<U, V, F: Fn(T, U) -> V>(self, other: XY<U>, f: F) -> XY<V>

Returns a new XY by calling f on self and other for each axis.

impl<T> XY<Option<T>>

fn unwrap_or(self, other: XY<T>) -> XY<T>

Returns a new XY by calling unwrap_or on each axis.

impl XY<bool>

fn any(&self) -> bool

Returns true if any of x or y is true.

fn both(&self) -> bool

Returns true if both x and y are true.

impl<T: Copy> XY<T>

fn both_from(value: T) -> Self

Creates a XY with both x and y set to value.