Enum cursive::theme::ColorStyle

pub enum ColorStyle {
    Background,
    Shadow,
    Primary,
    Secondary,
    Tertiary,
    TitlePrimary,
    TitleSecondary,
    Highlight,
    HighlightInactive,
}

Possible color style for a cell.

Represents a color pair role to use when printing something.

The current theme will assign each role a foreground and background color.

Variants
Background

Application background, where no view is present.

Shadow

Color used by view shadows. Only background matters.

Primary

Main text with default background.

Secondary

Secondary text color, with default background.

Tertiary

Tertiary text color, with default background.

TitlePrimary

Title text color with default background.

TitleSecondary

Alternative color for a title.

Highlight

Alternate text with highlight background.

HighlightInactive

Highlight color for inactive views (not in focus).

Methods

impl ColorStyle

fn id(self) -> i16

Returns the ncurses pair ID associated with this color pair.