Module cursive::utils

Toolbox to make text layout easier.

Structs

LinesIterator
Generates rows of text in constrained width.

ProgressReader
Wrapper around a Read that reports the progress made.

Row
Represents a row of text within a String.

Functions

prefix_length
Computes the length of a prefix that fits in the given width.

simple_suffix_length
Computes the length of a suffix that fits in the given width.

suffix_length
Computes the length of a suffix that fits in the given width.