Struct cursive::view::Dialog

pub struct Dialog {
    // some fields omitted
}

Popup-like view with a main content, and optional buttons under it.

Examples

let dialog = Dialog::new(TextView::new("Hello!")).button("Ok", |s| s.quit());

Methods

impl Dialog

fn new<V: View + 'static>(view: V) -> Self

Creates a new Dialog with the given content.

fn button<'a, F>(self, label: &'a str, cb: F) -> Self where F: Fn(&mut Cursive) + 'static

Adds a button to the dialog with the given label and callback.

Consumes and returns self for easy chaining.

fn h_align(self, h: HAlign) -> Self

Sets the horizontal alignment for the buttons, if any. Only works if the buttons are as a row at the bottom of the dialog.

fn v_align(self, v: VAlign) -> Self

Sets the vertical alignment for the buttons, if any. Only works if the buttons are as a column to the right of the dialog.

fn dismiss_button<'a>(self, label: &'a str) -> Self

Shortcut method to add a button that will dismiss the dialog.

fn title(self, label: &str) -> Self

Sets the title of the dialog. If not empty, it will be visible at the top.

fn padding<T: ToVec4>(self, padding: T) -> Self

Sets the padding in the dialog (around content and buttons).