Struct cursive::view::EditView

pub struct EditView {
    // some fields omitted
}

Input box where the user can enter and edit text.

Methods

impl EditView

fn new() -> Self

Creates a new, empty edit view.

fn set_content<'a>(&mut self, content: &'a str)

Replace the entire content of the view with the given one.

fn get_content(&self) -> &str

Get the current text.

fn content<'a>(self, content: &'a str) -> Self

Sets the current content to the given value. Convenient chainable method.

fn min_length(self, min_length: usize) -> Self

Sets the minimum length for this view. (This applies to the layout, not the content.)

fn with_id(self, label: &str) -> IdView<Self>

Wraps this view into an IdView with the given id.