Trait cursive::view::ViewWrapper

pub trait ViewWrapper {
    fn get_view(&self) -> &View;
    fn get_view_mut(&mut self) -> &mut View;

    fn wrap_draw(&mut self, printer: &Printer) { ... }
    fn wrap_get_min_size(&self, req: SizeRequest) -> Vec2 { ... }
    fn wrap_on_event(&mut self, ch: Event) -> EventResult { ... }
    fn wrap_layout(&mut self, size: Vec2) { ... }
    fn wrap_take_focus(&mut self) -> bool { ... }
    fn wrap_find(&mut self, selector: &Selector) -> Option<&mut Any> { ... }
}

Generic wrapper around a view.

Default implementation forwards all calls to the child view. Overrides some methods as desired.

Required Methods

fn get_view(&self) -> &View

Get an immutable reference to the wrapped view, so that we can forward some calls to it.

fn get_view_mut(&mut self) -> &mut View

Get a mutable reference to the wrapped view, for the mutable methods.

Provided Methods

fn wrap_draw(&mut self, printer: &Printer)

Wraps the draw method.

fn wrap_get_min_size(&self, req: SizeRequest) -> Vec2

Wraps the get_min_size method.

fn wrap_on_event(&mut self, ch: Event) -> EventResult

Wraps the on_event method.

fn wrap_layout(&mut self, size: Vec2)

Wraps the layout method

fn wrap_take_focus(&mut self) -> bool

Wraps the take_focus method

fn wrap_find(&mut self, selector: &Selector) -> Option<&mut Any>