Struct cursive::views::Button

pub struct Button { /* fields omitted */ }

Simple text label with a callback when is pressed.

A button shows its content in a single line and has a fixed size.

Examples

let quit_button = Button::new("Quit", |s| s.quit());

Methods

impl Button

pub fn new<F, S: Into<String>>(label: S, cb: F) -> Self where
F: Fn(&mut Cursive) + 'static,

Creates a new button with the given content and callback.

pub fn disable(&mut self)

Disables this view.

A disabled view cannot be selected.

pub fn disabled(self) -> Self

Disables this view.

Chainable variant.

pub fn enable(&mut self)

Re-enables this view.

pub fn set_enabled(&mut self, enabled: bool)

Enable or disable this view.

pub fn is_enabled(&self) -> bool

Returns true if this view is enabled.