Struct cursive::views::Checkbox

pub struct Checkbox { /* fields omitted */ }

Checkable box.

Methods

impl Checkbox

pub fn disable(&mut self)

Disables this view.

A disabled view cannot be selected.

pub fn disabled(self) -> Self

Disables this view.

Chainable variant.

pub fn enable(&mut self)

Re-enables this view.

pub fn set_enabled(&mut self, enabled: bool)

Enable or disable this view.

pub fn is_enabled(&self) -> bool

Returns true if this view is enabled.

pub fn new() -> Self

Creates a new, unchecked checkbox.

pub fn set_on_change<F: 'static + Fn(&mut Cursive, bool)>(
&mut self,
on_change: F
)

Sets a callback to be used when the state changes.

pub fn on_change<F: 'static + Fn(&mut Cursive, bool)>(
self,
on_change: F
) -> Self

Sets a callback to be used when the state changes.

Chainable variant.

pub fn toggle(&mut self) -> EventResult

Toggles the checkbox state.

pub fn check(&mut self) -> EventResult

Check the checkbox.

pub fn checked(self) -> Self

Check the checkbox.

Chainable variant.

pub fn is_checked(&self) -> bool

Returns true if the checkbox is checked.

pub fn uncheck(&mut self) -> EventResult

Uncheck the checkbox.

pub fn unchecked(self) -> Self

Uncheck the checkbox.

Chainable variant.

pub fn set_checked(&mut self, checked: bool) -> EventResult

Sets the checkbox state.