Struct cursive::views::KeyEventView

pub struct KeyEventView<T: View> { /* fields omitted */ }

A simple wrapper view that catches some ignored event from its child.

If the event doesn't have a corresponding callback, it will stay ignored.

Examples

let view = KeyEventView::new(TextView::new("This view has an event!"))
                        .register('q', |s| s.quit())
                        .register(event::Key::Esc, |s| s.quit());

Methods

impl<T: View> KeyEventView<T>

fn new(view: T) -> Self

Wraps the given view in a new KeyEventView.

fn register<F, E: Into<Event>>(self, event: E, cb: F) -> Self where F: Fn(&mut Cursive) + 'static

Registers a callback when the given key is ignored by the child.