Struct cursive::views::ListView

pub struct ListView { /* fields omitted */ }

Displays a scrollable list of elements.

Methods

impl ListView

pub fn new() -> Self

Creates a new, empty ListView.

pub fn add_child<V: View + 'static>(&mut self, label: &str, view: V)

Adds a view to the end of the list.

pub fn clear(&mut self)

Removes all children from this view.

pub fn child<V: View + 'static>(self, label: &str, view: V) -> Self

Adds a view to the end of the list.

Chainable variant.

pub fn add_delimiter(&mut self)

Adds a delimiter to the end of the list.

pub fn delimiter(self) -> Self

Adds a delimiter to the end of the list.

Chainable variant.

pub fn set_on_select<F>(&mut self, cb: F) where
F: Fn(&mut Cursive, &String) + 'static,

Sets a callback to be used when an item is selected.

pub fn on_select<F>(self, cb: F) -> Self where
F: Fn(&mut Cursive, &String) + 'static,

Sets a callback to be used when an item is selected.

Chainable variant.

pub fn focus(&self) -> usize

Returns the index of the currently focused item.

Panics if the list is empty.