Struct cursive::views::Menubar

pub struct Menubar {
    pub autohide: bool,
    // some fields omitted
}

Shows a single-line list of items, with pop-up menus when one is selected.

The Cursive root already includes a menubar that you just need to configure.

Fields
autohide: bool

TODO: move this out of this view.

Methods

impl Menubar

fn new() -> Self

Creates a new, empty menubar.

fn receive_events(&self) -> bool

True if we should be receiving events.

fn visible(&self) -> bool

Returns true if we should be drawn.

fn add_subtree(&mut self, title: &str, menu: MenuTree) -> &mut Self

Adds a new item to the menubar.

The item will use the given title, and on selection, will open a popup-menu with the given menu tree.

fn insert_subtree(&mut self, i: usize, title: &str, menu: MenuTree) -> &mut Self

Insert a new item at the given position.

fn clear(&mut self)

Removes all menu items from this menubar.

fn len(&self) -> usize

Returns the number of items in this menubar.

fn is_empty(&self) -> bool

Returns true if this menubar is empty.

fn get_subtree(&mut self, i: usize) -> Option<&mut MenuTree>

Returns the item at the given position.

Returns None if i > self.len()

fn find_subtree(&mut self, label: &str) -> Option<&mut MenuTree>

Looks for an item with the given label.

fn find_position(&mut self, label: &str) -> Option<usize>

Returns the position of the item with the given label.

Returns None if no such label was found.

fn remove(&mut self, i: usize)

Remove the item at the given position.