Struct cursive::views::RadioButton

pub struct RadioButton<T> { /* fields omitted */ }

Variant of Checkbox arranged in group.

RadioButtons are managed by a RadioGroup. A single group can contain several radio buttons, but only one button per group can be active at a time.

RadioButtons are not created directly, but through RadioGroup::button().

Methods

impl<T> RadioButton<T>

fn disable(&mut self)

Disables this view.

A disabled view cannot be selected.

fn disabled(self) -> Self

Disables this view.

Chainable variant.

fn enable(&mut self)

Re-enables this view.

fn set_enabled(&mut self, enabled: bool)

Enable or disable this view.

fn is_enabled(&self) -> bool

Returns true if this view is enabled.

fn is_selected(&self) -> bool

Returns true if this button is selected.

fn select(&mut self)

Selects this button, un-selecting any other in the same group.

fn selected(self) -> Self

Selects this button, un-selecting any other in the same group.

Chainable variant.