Struct cursive::views::RadioGroup

pub struct RadioGroup<T> { /* fields omitted */ }

Group to coordinate multiple radio buttons.

A RadioGroup is used to create and manage RadioButtons.

A RadioGroup can be cloned; it will keep pointing to the same group.

Methods

impl<T> RadioGroup<T>

fn new() -> Self

Creates an empty group for radio buttons.

fn button<S: Into<String>>(&mut self, value: T, label: S) -> RadioButton<T>

Adds a new button to the group.

The button will display label next to it, and will embed value.

fn selected_id(&self) -> usize

Returns the id of the selected button.

Buttons are indexed in the order they are created, starting from 0.

fn selection(&self) -> Rc<T>

Returns the value associated with the selected button.

impl RadioGroup<String>

fn button_str<S: Into<String>>(&mut self, text: S) -> RadioButton<String>

Adds a button, using the label itself as value.