Struct cursive::views::StackView

pub struct StackView { /* fields omitted */ }

Simple stack of views. Only the top-most view is active and can receive input.

Methods

impl StackView

pub fn new() -> Self

Creates a new empty StackView

pub fn add_layer<T: 'static + View>(&mut self, view: T)

Adds new view on top of the stack in the center of the screen.

pub fn add_layer_at<T: 'static + View>(&mut self, position: Position, view: T)

Adds a view on top of the stack.

pub fn pop_layer(&mut self)

Remove the top-most layer.

pub fn offset(&self) -> Vec2

Computes the offset of the current top view.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn layer_sizes(&self) -> Vec<Vec2>

Returns the size for each layer in this view.