Crate cursive

Cursive

Cursive is a TUI library built on top of ncurses-rs. It allows to easily build layouts for text-based applications.

Example

extern crate cursive;

use cursive::Cursive;
use cursive::view::TextView;

fn main() {
    let mut siv = Cursive::new();

    siv.add_layer(TextView::new("Hello World!\nPress q to quit."));

    siv.add_global_callback('q', |s| s.quit());

    siv.run();
}

Modules

align
Tools to control view alignment

event
User-input events and their effects.

orientation
Define an Orientation and associated methods.

printer
Makes drawing on ncurses windows easier.

theme
Module to handle colors and themes in the UI.

vec
Points on the 2D character grid.

view
Defines various views to use when creating the layout.

Macros

wrap_impl!
Convenient macro to implement to two methods required for the ViewWrapper trait.

Structs

Cursive
Central part of the cursive library.

Type Definitions

ScreenId
Identifies a screen in the cursive ROOT.