Struct cursive::align::Align

pub struct Align {
    pub h: HAlign,
    pub v: VAlign,
}

Specifies the alignment along both horizontal and vertical directions.

Fields
h: HAlign

Horizontal alignment policy

v: VAlign

Vertical alignment policy

Methods

impl Align

fn new(h: HAlign, v: VAlign) -> Self

Creates a new Align object from the given alignments.

fn top_left() -> Self

Creates a top-left alignment.

fn top_right() -> Self

Creates a top-right alignment.

fn bot_left() -> Self

Creates a bottom-left alignment.

fn bot_right() -> Self

Creates a bottom-right alignment.

fn center() -> Self

Creates an alignment centered both horizontally and vertically.