Enum cursive::event::Event

pub enum Event {
    WindowResize,
    Refresh,
    Char(char),
    CtrlChar(char),
    AltChar(char),
    Key(Key),
    Shift(Key),
    Alt(Key),
    AltShift(Key),
    Ctrl(Key),
    CtrlShift(Key),
    CtrlAlt(Key),
    Unknown(i32),
}

Represents an event as seen by the application.

Variants
WindowResize

Event fired when the window is resized.

Refresh

Event fired regularly when a auto-refresh is set.

Char(char)

A character was entered (includes numbers, punctuation, ...).

CtrlChar(char)

A character was entered with the Ctrl key pressed.

AltChar(char)

A character was entered with the Alt key pressed.

Key(Key)

A non-character key was pressed.

Shift(Key)

A non-character key was pressed with the Shift key pressed.

Alt(Key)

A non-character key was pressed with the Alt key pressed.

AltShift(Key)

A non-character key was pressed with the Shift and Alt keys pressed.

Ctrl(Key)

A non-character key was pressed with the Ctrl key pressed.

CtrlShift(Key)

A non-character key was pressed with the Ctrl and Shift keys pressed.

CtrlAlt(Key)

A non-character key was pressed with the Ctrl and Alt keys pressed.

Unknown(i32)

An unknown event was received.