Enum cursive::event::Key

pub enum Key {
    Enter,
    Tab,
    Backspace,
    Esc,
    Left,
    Right,
    Up,
    Down,
    Ins,
    Del,
    Home,
    End,
    PageUp,
    PageDown,
    PauseBreak,
    NumpadCenter,
    F0,
    F1,
    F2,
    F3,
    F4,
    F5,
    F6,
    F7,
    F8,
    F9,
    F10,
    F11,
    F12,
}

A non-character key on the keyboard

Variants
Enter

Both Enter (or Return) and numpad Enter

Tab

Tabulation key

Backspace

Backspace key

Esc

Escape key

Left

Left arrow

Right

Right arrow

Up

Up arrow

Down

Down arrow

Ins

Insert key

Del

Delete key

Home

Home key

End

End key

PageUp

Page Up key

PageDown

Page Down key

PauseBreak

Pause Break key

NumpadCenter

The 5 in the center of the keypad, when numlock is disabled.

F0

F0 key

F1

F1 key

F2

F2 key

F3

F3 key

F4

F4 key

F5

F5 key

F6

F6 key

F7

F7 key

F8

F8 key

F9

F9 key

F10

F10 key

F11

F11 key

F12

F12 key

Methods

impl Key

pub fn from_f(n: u8) -> Key

Returns the function key corresponding to the given number

1 -> F1, etc...

Panics

If n == 0 || n > 12