Struct cursive::XY

pub struct XY<T> {
    pub x: T,
    pub y: T,
}

A generic structure with a value for each axis.

Fields
x: T

X-axis value

y: T

Y-axis value

Methods

impl<T: Ord> XY<T>

pub fn fits_in<O: Into<Self>>(&self, other: O) -> bool

Returns true if self could fit inside other.

Shortcut for self.x <= other.x && self.y <= other.y.

pub fn max<A: Into<XY<T>>, B: Into<XY<T>>>(a: A, b: B) -> Self

Returns a new Vec2 that is a maximum per coordinate.

pub fn min<A: Into<XY<T>>, B: Into<XY<T>>>(a: A, b: B) -> Self

Returns a new Vec2 that is no larger than any input in both dimensions.

pub fn or_min<O: Into<XY<T>>>(self, other: O) -> Self

Returns the minimum of self and other.

pub fn or_max<O: Into<XY<T>>>(self, other: O) -> Self

Returns the maximum of self and other.

impl<T: Ord + Add<Output = T> + Clone> XY<T>

pub fn stack_vertical(&self, other: &Self) -> Self

Returns (max(self.x,other.x), self.y+other.y)

pub fn stack_horizontal(&self, other: &Self) -> Self

Returns (self.x+other.x, max(self.y,other.y))

impl<T: Zero + Clone> XY<T>

pub fn keep_x(&self) -> Self

Returns a vector with the X component of self, and y=0.

pub fn keep_y(&self) -> Self

Returns a vector with the Y component of self, and x=0.

pub fn zero() -> Self

Alias for Self::new(0,0).

impl<T> XY<T>

pub fn new(x: T, y: T) -> Self

Creates a new XY from the given values.

pub fn map<U, F: Fn(T) -> U>(self, f: F) -> XY<U>

Creates a new XY by applying f to x and y.

pub fn pair(self) -> (T, T)

Destructure self into a pair.

pub fn as_ref(&self) -> XY<&T>

Return a XY with references to this one's values.

Important traits for Chain<A, B>
impl<A, B> Iterator for Chain<A, B> where
A: Iterator,
B: Iterator<Item = <A as Iterator>::Item>, type Item = <A as Iterator>::Item;

pub fn iter(&self) -> Chain<Once<&T>, Once<&T>>

Creates an iterator that returns references to x, then y.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get(&self, o: Orientation) -> &T

Returns a reference to the value on the given axis.

pub fn zip<U>(self, other: XY<U>) -> XY<(T, U)>

Returns a new XY of tuples made by zipping self and other.

pub fn zip_map<U, V, F: Fn(T, U) -> V>(self, other: XY<U>, f: F) -> XY<V>

Returns a new XY by calling f on self and other for each axis.

impl<T: Clone> XY<T>

pub fn with_axis(&self, o: Orientation, value: T) -> Self

Returns a new XY with the axis o set to value.

pub fn with_axis_from(&self, o: Orientation, other: &Self) -> Self

Returns a new XY with the axis o set to the value from other.

pub fn set_axis_from(&mut self, o: Orientation, other: &Self)

Sets the axis o on self to the value from other.

impl<T> XY<Option<T>>

pub fn unwrap_or(self, other: XY<T>) -> XY<T>

Returns a new XY by calling unwrap_or on each axis.

impl XY<bool>

pub fn any(&self) -> bool

Returns true if any of x or y is true.

pub fn both(&self) -> bool

Returns true if both x and y are true.

impl<T: Copy> XY<T>

pub fn both_from(value: T) -> Self

Creates a XY with both x and y set to value.