Module cursive::theme

Module to handle colors and themes in the UI.

Structs

Palette
Color configuration for the application.

Theme
Represents the style a Cursive application will use.

Enums

BorderStyle
Specifies how some borders should be drawn.

Color
Represents a color used by the theme.

ColorStyle
Possible color style for a cell.

Effect

Error
Possible error returned when loading a theme.

Functions

load_default
Loads the default theme, and returns its representation.

load_theme
Loads a theme file, and returns its representation if everything worked well.