Module cursive::view

Defines various views to use when creating the layout.

Structs

BoxView
BoxView is a wrapper around an other view, with a given minimum size.

Button
Simple text label with a callback when ENTER is pressed. A button shows its content in a single line and has a fixed size.

Dialog
Popup-like view with a main content, and optional buttons under it.

EditView
Input box where the user can enter and edit text.

FullView
Simple wrapper view that asks for all the space it can get.

IdView
Wrapper view that allows to select its content with a fixed string id.

KeyEventView
A simple wrapper view that catches some ignored event from its child.

LinearLayout
Arranges its children linearly according to its orientation.

MenuPopup
fd

Position
Location of the view on screen

ScrollBase
Provide scrolling functionalities to a view.

SelectView
View to select an item among a list.

ShadowView
Wrapper view that adds a shadow.

SizedView
Wrapper around a view that remembers its size.

StackView
Simple stack of views. Only the top-most view is active and can receive input.

TextView
A simple view showing a fixed text

TrackedView

ViewPath
Represents a path to a single view in the layout.

Enums

Offset

Selector
Selects a single view (if any) in the tree.

Traits

View
Main trait defining a view behaviour.

ViewWrapper
Generic wrapper around a view.