Module cursive::views

Various views to use when creating the layout.

Structs

BoxView
Wrapper around another view, with a controlled size.

Button
Simple text label with a callback when is pressed.

Checkbox
Checkable box.

Counter
Atomic counter used by ProgressBar.

Dialog
Popup-like view with a main content, and optional buttons under it.

DummyView
Dummy view.

EditView
Input box where the user can enter and edit text.

IdView
Wrapper view that allows to select its content with a fixed string id.

KeyEventView
A simple wrapper view that catches some ignored event from its child.

LinearLayout
Arranges its children linearly according to its orientation.

ListView
Displays a scrollable list of elements.

MenuPopup
Popup that shows a list of items.

Menubar
Shows a single-line list of items, with pop-up menus when one is selected.

Panel
Draws a border around a wrapped view.

ProgressBar
Animated bar showing a progress value.

RadioButton
Variant of Checkbox arranged in group.

RadioGroup
Group to coordinate multiple radio buttons.

SelectView
View to select an item among a list.

ShadowView
Wrapper view that adds a shadow.

SizedView
Wrapper around a view that remembers its size.

SliderView
A horizontal or vertical slider.

StackView
Simple stack of views. Only the top-most view is active and can receive input.

TextArea
Multi-lines text editor.

TextView
A simple view showing a fixed text

TrackedView
Wrapper around a view that remembers its position.