Struct cursive::views::EditView

pub struct EditView { /* fields omitted */ }

Input box where the user can enter and edit text.

Examples

From the edit example.

let mut siv = Cursive::new();

// Create a dialog with an edit text and a button.
// The user can either hit the <Ok> button,
// or press Enter on the edit text.
siv.add_layer(Dialog::new()
    .title("Enter your name")
    .padding((1, 1, 1, 0))
    .content(EditView::new()
        .on_submit(show_popup)
        .with_id("name")
        .fixed_width(20))
    .button("Ok", |s| {
        let name = s.find_id::<EditView>("name")
            .unwrap()
            .get_content();
        show_popup(s, &name);
    }));

fn show_popup(s: &mut Cursive, name: &str) {
    if name.is_empty() {
        s.add_layer(Dialog::info("Please enter a name!"));
    } else {
        let content = format!("Hello {}!", name);
        s.pop_layer();
        s.add_layer(Dialog::around(TextView::new(content))
            .button("Quit", |s| s.quit()));
    }
}

Methods

impl EditView

pub fn new() -> Self

Creates a new, empty edit view.

pub fn set_secret(&mut self, secret: bool)

If secret is true, the content won't be displayed in clear.

Only * will be shown.

pub fn secret(self) -> Self

Hides the content of the view.

Only * will be shown.

pub fn disable(&mut self)

Disables this view.

A disabled view cannot be selected.

pub fn disabled(self) -> Self

Disables this view.

Chainable variant.

pub fn enable(&mut self)

Re-enables this view.

pub fn on_edit<F>(self, callback: F) -> Self where
F: Fn(&mut Cursive, &str, usize) + 'static,

Sets a callback to be called whenever the content is modified.

callback will be called with the view content and the current cursor position.

pub fn on_submit<F>(self, callback: F) -> Self where
F: Fn(&mut Cursive, &str) + 'static,

Sets a callback to be called when <Enter> is pressed.

callback will be given the content of the view.

pub fn set_enabled(&mut self, enabled: bool)

Enable or disable this view.

pub fn is_enabled(&self) -> bool

Returns true if this view is enabled.

pub fn set_content<S: Into<String>>(&mut self, content: S)

Replace the entire content of the view with the given one.

pub fn get_content(&self) -> Rc<String>

Get the current text.

pub fn content<S: Into<String>>(self, content: S) -> Self

Sets the current content to the given value.

Convenient chainable method.

pub fn set_cursor(&mut self, cursor: usize)

Sets the cursor position.

pub fn insert(&mut self, ch: char)

Insert ch at the current cursor position.

pub fn remove(&mut self, len: usize)

Remove the character at the current cursor position.