Struct cursive::views::ListView

pub struct ListView {
    // some fields omitted
}

Displays a scrollable list of elements.

Methods

impl ListView

fn new() -> Self

Creates a new, empty ListView.

fn add_child<V: View + 'static>(&mut self, label: &str, view: V)

Adds a view to the end of the list.

fn clear(&mut self)

Removes all children from this view.

fn child<V: View + 'static>(self, label: &str, view: V) -> Self

Adds a view to the end of the list.

Chainable variant.

fn add_delimiter(&mut self)

Adds a delimiter to the end of the list.

fn delimiter(self) -> Self

Adds a delimiter to the end of the list.

Chainable variant.