Struct cursive::views::SliderView

pub struct SliderView { /* fields omitted */ }

A horizontal or vertical slider.

Methods

impl SliderView

fn new(orientation: Orientation, max_value: usize) -> Self

Creates a new SliderView in the given orientation.

The view will have a fixed length of max_value, with one tick per block.

fn vertical(max_value: usize) -> Self

Creates a new vertical SliderView.

fn horizontal(max_value: usize) -> Self

Creates a new horizontal SliderView.

fn set_value(&mut self, value: usize) -> EventResult

Sets the current value.

Returns an event result with a possible callback, if on_change was set..

fn value(self, value: usize) -> Self

Sets the current value.

Chainable variant.

fn on_change<F>(self, callback: F) -> Self where F: Fn(&mut Cursive, usize) + 'static

Sets a callback to be called when the slider is moved.

fn on_enter<F>(self, callback: F) -> Self where F: Fn(&mut Cursive, usize) + 'static

Sets a callback to be called when the key is pressed.