Struct cursive::views::TextView

pub struct TextView { /* fields omitted */ }

A simple view showing a fixed text

Methods

impl TextView

fn new<S: Into<String>>(content: S) -> Self

Creates a new TextView with the given content.

fn set_scrollable(&mut self, scrollable: bool)

Enable or disable the view's scrolling capabilities.

When disabled, the view will never attempt to scroll (and will always ask for the full height).

fn scrollable(self, scrollable: bool) -> Self

Enable or disable the view's scrolling capabilities.

When disabled, the view will never attempt to scroll (and will always ask for the full height).

Chainable variant.

fn h_align(self, h: HAlign) -> Self

Sets the horizontal alignment for this view.

fn v_align(self, v: VAlign) -> Self

Sets the vertical alignment for this view.

fn align(self, a: Align) -> Self

Sets the alignment for this view.

fn center(self) -> Self

Center the text horizontally and vertically inside the view.

fn set_content<S: Into<String>>(&mut self, content: S)

Replace the text in this view.

fn get_content(&self) -> &str

Returns the current text in this view.